resource를 view에서 받아 사용하도록 수정
