+++ app/src/main/java/com/yeonae/chamelezone/view/review/MyReviewDetailActivity.kt
@@ -7,6 +7,14 @@ import com.yeonae.chamelezone.view.review.adapter.ReviewImageVpAdapter
 import kotlinx.android.synthetic.main.activity_myreview_detail.*
 
 class MyReviewDetailActivity : AppCompatActivity() {
+
+    val images = intArrayOf(
+        R.drawable.img1,
+        R.drawable.img2,
+        R.drawable.img3,
+        R.drawable.img4
+    )
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_myreview_detail)
